Sat, Jul 6, 2024, 6:40 PM +00:00

+++ tip/tip.py
@@ -6,7 +6,7 @@ def main():
 
 
 def dollars_to_float(d):
-    d = int(dollars.replace("$",""))
+    d = float(dollars.replace("$",""))
     return d
 
 def percent_to_float(p):
